requirements-dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements-dev.txt requirements-dev.in
#
autopep8==1.3.1
click==6.7                # via pip-tools
colorama==0.4.4           # via pytest
configparser==3.5.0       # via flake8
coverage==4.3.4           # via pytest-cov
enum34==1.1.6             # via flake8
first==2.0.1              # via pip-tools
flake8==3.3.0
freezegun==0.3.15
isort==4.2.5
mccabe==0.6.1             # via flake8
pep8==1.7.0
pip-tools==1.9.0
py==1.4.33                # via pytest
pycodestyle==2.3.1        # via autopep8, flake8
pyflakes==1.5.0           # via flake8
pytest-cov==2.4.0
pytest-django==3.1.2
pytest==3.0.7
python-dateutil==2.8.1    # via freezegun
six==1.15.0               # via freezegun, pip-tools, python-dateutil
